+++ composeApp/src/wasmJsMain/kotlin/org/dotnet/app/CarRentalApp.kt
@@ -1,7 +1,7 @@
 package org.dotnet.app
 
 import androidx.compose.runtime.*
-import org.dotnet.app.view.RentCarScreen
+import org.dotnet.app.presentation.screens.RentCarScreen
 import org.dotnet.app.viewModel.CarRentalAppViewModel
 
 @Composable
