+++ composeApp/src/wasmJsMain/kotlin/org/dotnet/app/domain/AuthResponse.kt
@@ -1,4 +1,4 @@
-package org.dotnet.app.model
+package org.dotnet.app.domain
 
 import kotlinx.serialization.Serializable
