+++ composeApp/src/wasmJsMain/kotlin/org/dotnet/app/domain/CredentialResponse.kt
@@ -1,4 +1,4 @@
-package org.dotnet.app.model
+package org.dotnet.app.domain
 
 data class CredentialResponse(
     var credential: String
