+++ composeApp/src/wasmJsMain/kotlin/org/dotnet/app/main.kt
@@ -4,7 +4,6 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
-import org.dotnet.app.view.RentCarScreen
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
