+++ composeApp/src/wasmJsMain/kotlin/org/dotnet/app/presentation/screens/ValuationScreen.kt
@@ -1,9 +1,7 @@
-package org.dotnet.app.view
+package org.dotnet.app.presentation.screens
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
@@ -17,7 +15,7 @@ import androidx.compose.ui.Alignment
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
-import org.dotnet.app.model.Car
+import org.dotnet.app.domain.Car
 
 
 @Composable
